5_29

+++ GPSR Web development/GPSR/kgg_GPSR/GPSR.py
@@ -342,38 +342,66 @@ def talk_change():
 
 @app.route('/geonguprincesssecretroom')
 def admin():
+    if not g.user:
+        return redirect(url_for('home'))
+    if g.user['user_id']!='admin':
+        return redirect(url_for('home'))
     return render_template('/admin/admin.html')
 
 #view user_list
 @app.route('/geonguprincesssecretroom/view_user')
 def admin_view_user():
+    if not g.user:
+        return redirect(url_for('home'))
+    if g.user['user_id']!='admin':
+        return redirect(url_for('home'))
     return render_template('/admin/admin_view_user.html',users=query_db('''
     select * from user limit ?''', [PER_PAGE]))
     
 @app.route('/geonguprincesssecretroom/view_problem')
 def admin_view_problem():
+    if not g.user:
+        return redirect(url_for('home'))
+    if g.user['user_id']!='admin':
+        return redirect(url_for('home'))
     return render_template('/admin/admin_view_problem.html',problems=query_db('''
     select * from problem limit?''', [PER_PAGE]))
 
 #all of user information
 @app.route('/geonguprincesssecretroom/view_user_info/<user_id>')
 def admin_view_user_info(user_id):
+    if not g.user:
+        return redirect(url_for('home'))
+    if g.user['user_id']!='admin':
+        return redirect(url_for('home'))
     user = query_db('select * from user where user_id = ?', [user_id], True)
     return render_template('/admin/admin_view_user_info.html', user=user)
 
 #user delete
 @app.route('/geonguprincesssecretroom/delete_user/<user_id>')
 def admin_delete_user(user_id):
+    if not g.user:
+        return redirect(url_for('home'))
+    if g.user['user_id']!='admin':
+        return redirect(url_for('home'))
     g.db.execute('delete from user where user_id = ?', [user_id])
     g.db.commit()
     return redirect(url_for('admin_view_user'))
 
 @app.route('/geonguprincesssecretroom/add_problem')
 def admin_add_problem():
+    if not g.user:
+        return redirect(url_for('home'))
+    if g.user['user_id']!='admin':
+        return redirect(url_for('home'))
     return render_template('/admin/admin_add_problem.html', error=None)
 
 @app.route('/geonguprincesssecretroom/add_problem_exe',methods=['POST'])
 def admin_add_problem_exe():
+    if not g.user:
+        return redirect(url_for('home'))
+    if g.user['user_id']!='admin':
+        return redirect(url_for('home'))
     error=None
     if request.method == 'POST':
         if not request.form['problem_name']:
@@ -393,22 +421,38 @@ def admin_add_problem_exe():
 
 @app.route('/geonguprincesssecretroom/delete_problem/<problem_num>')
 def admin_delete_problem(problem_num):
+    if not g.user:
+        return redirect(url_for('home'))
+    if g.user['user_id']!='admin':
+        return redirect(url_for('home'))
     g.db.execute('delete from problem where problem_num = ?', [problem_num])
     g.db.commit()
     return redirect(url_for('admin_view_problem'))
 
 @app.route('/geonguprincesssecretroom/view_problem_info/<problem_num>',methods=['POST','GET'])
 def admin_view_problem_info(problem_num):
+    if not g.user:
+        return redirect(url_for('home'))
+    if g.user['user_id']!='admin':
+        return redirect(url_for('home'))
     problem = query_db('select * from problem where problem_num = ?', [problem_num], True)
     return render_template('/admin/admin_view_problem_info.html', problem=problem)
 
 @app.route('/geonguprincesssecretroom/problem_change/<problem_num>',methods=['GET'])
 def admin_problem_change_view(problem_num):
+    if not g.user:
+        return redirect(url_for('home'))
+    if g.user['user_id']!='admin':
+        return redirect(url_for('home'))
     problem = query_db('select * from problem where problem_num = ?', [problem_num])
     return render_template('/admin/admin_view_problem_change.html',problem=problem)
 
 @app.route('/geonguprincesssecretroom/problem_changing',methods=['POST'])
 def admin_problem_changing():
+    if not g.user:
+        return redirect(url_for('home'))
+    if g.user['user_id']!='admin':
+        return redirect(url_for('home'))
     g.db.execute('''update problem set problem_name = ? , problem_text_info = ? , problem_text_input_info = ? , 
     problem_text_output_info = ? where problem_num = ? '''
                  ,[request.form['problem_title'], request.form['problem_body'], request.form['problem_input'], request.form['problem_output'] ,request.form['num']])
@@ -417,6 +461,8 @@ def admin_problem_changing():
 
 @app.route('/geonguprincesssecretroom/view_talk')
 def admin_view_talk():
+    if not g.user:
+        return redirect(url_for('home'))
     if g.user['user_id']!='admin':
         return redirect(url_for('home'))
     talk = query_db('select * from board')
@@ -424,6 +470,8 @@ def admin_view_talk():
 
 @app.route('/geonguprincesssecretroom/view_talk/<board_num>',methods=['GET','POST'])
 def admin_view_talk_more(board_num):
+    if not g.user:
+        return redirect(url_for('home'))
     if g.user['user_id']!='admin':
         return redirect(url_for('home'))
     talk = query_db('select * from board where board_num is ? ', [board_num])
@@ -431,6 +479,8 @@ def admin_view_talk_more(board_num):
 
 @app.route('/geonguprincesssecretroom/talk_delete/<board_num>',methods=['POST','GET'])
 def admin_talk_delete(board_num):
+    if not g.user:
+        return redirect(url_for('home'))
     if g.user['user_id']!='admin':
         return redirect(url_for('home'))
     g.db.execute('delete from board where board_num = ?', board_num)
@@ -439,6 +489,8 @@ def admin_talk_delete(board_num):
 
 @app.route('/geonguprincesssecretroom/talk_change/<board_num>',methods=['POST','GET'])
 def admin_talk_change_view(board_num):
+    if not g.user:
+        return redirect(url_for('home'))
     if g.user['user_id']!='admin':
         return redirect(url_for('home'))
     talk = query_db('select * from board where board_num is ?',[board_num])
@@ -446,6 +498,8 @@ def admin_talk_change_view(board_num):
 
 @app.route('/geonguprincesssecretroom/talk_changing', methods=['GET','POST'])
 def admin_talk_change():
+    if not g.user:
+        return redirect(url_for('home'))
     if g.user['user_id']!='admin':
         return redirect(url_for('home'))
     g.db.execute('update board set board_name = ? , board_text = ? where board_num = ? '
